# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type UserSubscriptionStatus = typeof UserSubscriptionStatus[keyof typeof UserSubscriptionStatus];


export const UserSubscriptionStatus = {
  none: 'none',
  pending: 'pending',
  active: 'active',
  rejected: 'rejected',
  expired: 'expired',
} as const;

export interface User {
  id: string;
  name: string;
  email: string;
  isAdmin: boolean;
  subscriptionStatus: UserSubscriptionStatus;
}

export interface Session {
  user: User;
}

export interface LoginInput {
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export type ContentSectionType = typeof ContentSectionType[keyof typeof ContentSectionType];


export const ContentSectionType = {
  explanations: 'explanations',
  homework: 'homework',
  exams: 'exams',
  qiyas: 'qiyas',
} as const;

export interface ContentSection {
  id: string;
  type: ContentSectionType;
  title: string;
  count: number;
}

export interface Subject {
  id: string;
  name: string;
  icon: string;
  lessonCount: number;
  sections: ContentSection[];
}

export interface Grade {
  id: string;
  name: string;
  subjects: Subject[];
}

export interface Stage {
  id: string;
  name: string;
  description: string;
  color: string;
  grades: Grade[];
}

export interface Catalog {
  stages: Stage[];
}

export interface Lesson {
  id: string;
  title: string;
  subjectName: string;
  content: string;
  locked: boolean;
}

export type SubscriptionStatus = typeof SubscriptionStatus[keyof typeof SubscriptionStatus];


export const SubscriptionStatus = {
  none: 'none',
  pending: 'pending',
  active: 'active',
  rejected: 'rejected',
  expired: 'expired',
} as const;

export type SubscriptionPlan = typeof SubscriptionPlan[keyof typeof SubscriptionPlan];


export const SubscriptionPlan = {
  monthly: 'monthly',
  yearly: 'yearly',
} as const;

export interface Subscription {
  status: SubscriptionStatus;
  plan: SubscriptionPlan;
  amount: number;
  /** @nullable */
  receiptName?: string | null;
  /** @nullable */
  submittedAt: string | null;
}

export type SubscriptionInputPlan = typeof SubscriptionInputPlan[keyof typeof SubscriptionInputPlan];


export const SubscriptionInputPlan = {
  monthly: 'monthly',
  yearly: 'yearly',
} as const;

export interface SubscriptionInput {
  plan: SubscriptionInputPlan;
  receiptName: string;
}

export type SubscriptionRequest = Subscription & {
  id: string;
  user: User;
};

export type SubscriptionStatusInputStatus = typeof SubscriptionStatusInputStatus[keyof typeof SubscriptionStatusInputStatus];


export const SubscriptionStatusInputStatus = {
  active: 'active',
  rejected: 'rejected',
} as const;

export interface SubscriptionStatusInput {
  status: SubscriptionStatusInputStatus;
}

export interface AdminOverview {
  students: number;
  activeSubscriptions: number;
  pendingRequests: number;
  lessons: number;
}

export type CatalogItemInputType = typeof CatalogItemInputType[keyof typeof CatalogItemInputType];


export const CatalogItemInputType = {
  stage: 'stage',
  grade: 'grade',
  subject: 'subject',
  lesson: 'lesson',
} as const;

export interface CatalogItemInput {
  type: CatalogItemInputType;
  name: string;
  /** @nullable */
  parentId?: string | null;
}

export type CatalogItem = CatalogItemInput & {
  id: string;
};

export interface Settings {
  monthlyPrice: number;
  yearlyPrice: number;
  bankName: string;
  accountName: string;
  iban: string;
}

export type SettingsInput = Settings;

export type AiGenerationInputType = typeof AiGenerationInputType[keyof typeof AiGenerationInputType];


export const AiGenerationInputType = {
  lesson: 'lesson',
  homework: 'homework',
  exam: 'exam',
  qiyas: 'qiyas',
} as const;

export interface AiGenerationInput {
  type: AiGenerationInputType;
  subject: string;
  grade: string;
  topic?: string;
  /**
     * Optional transient image data URL for solving a photographed homework question
     * @maxLength 7000000
     */
  imageData?: string;
  /**
     * Selects one of the 20 local fallback tests when the AI provider is unavailable
     * @minimum 1
     * @maximum 20
     */
  fallbackTestNumber?: number;
}

export type AiGenerationResultItemsItemDifficulty = typeof AiGenerationResultItemsItemDifficulty[keyof typeof AiGenerationResultItemsItemDifficulty];


export const AiGenerationResultItemsItemDifficulty = {
  سهل: 'سهل',
  متوسط: 'متوسط',
  صعب: 'صعب',
} as const;

export type AiGenerationResultItemsItem = {
  question: string;
  answer: string;
  explanation: string;
  /** Multiple-choice options for a generated practice question */
  options?: string[];
  skill?: string;
  difficulty?: AiGenerationResultItemsItemDifficulty;
};

export type AiGenerationResultSource = typeof AiGenerationResultSource[keyof typeof AiGenerationResultSource];


export const AiGenerationResultSource = {
  ai: 'ai',
  fallback: 'fallback',
} as const;

export interface AiGenerationResult {
  title: string;
  body: string;
  items: AiGenerationResultItemsItem[];
  source?: AiGenerationResultSource;
  testNumber?: number;
  testsAvailable?: number;
  questionsPerTest?: number;
}