# lib/api-zod/src/generated/types/aiGenerationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiGenerationInputType } from './aiGenerationInputType';

export interface AiGenerationInput {
  type: AiGenerationInputType;
  subject: string;
  grade: string;
  topic?: string;
  /**
     * Optional transient image data URL for solving a photographed homework question
     * @maxLength 7000000
     */
  imageData?: string;
  /**
     * Selects one of the 20 local fallback tests when the AI provider is unavailable
     * @minimum 1
     * @maximum 20
     */
  fallbackTestNumber?: number;
}