# lib/api-zod/src/generated/types/aiGenerationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AiGenerationResultItemsItem } from './aiGenerationResultItemsItem';
import type { AiGenerationResultSource } from './aiGenerationResultSource';

export interface AiGenerationResult {
  title: string;
  body: string;
  items: AiGenerationResultItemsItem[];
  source?: AiGenerationResultSource;
  testNumber?: number;
  testsAvailable?: number;
  questionsPerTest?: number;
}