# lib/api-zod/src/generated/types/catalog.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Stage } from './stage';

export interface Catalog {
  stages: Stage[];
}