# lib/api-zod/src/generated/types/catalogItemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CatalogItemInputType } from './catalogItemInputType';

export interface CatalogItemInput {
  type: CatalogItemInputType;
  name: string;
  /** @nullable */
  parentId?: string | null;
}