# lib/api-zod/src/generated/types/grade.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Subject } from './subject';

export interface Grade {
  id: string;
  name: string;
  subjects: Subject[];
}