# lib/api-zod/src/generated/types/lesson.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lesson {
  id: string;
  title: string;
  subjectName: string;
  content: string;
  locked: boolean;
}